% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{examplehemi}
\alias{examplehemi}
\alias{largegap}
\alias{smallgap}
\title{Two example hemiphotos in YplantQMC format}
\source{
Thanks to Daniel Falster and Bob Pearcy.
}
\description{
A canopy with a small gap, and a canopy with a large gap. Both are objects of
class \code{yphemi}, which is normally constructed with the function
\code{\link{setHemi}}.
}
\examples{
\dontrun{
# Small gap
plot(smallgap)

# Large gap, with solar path in June in the south of france.
southfrance <- setLocation(44)
juneday <- setMet(southfrance, month=6, day=21)
plot(largegap, juneday)
}
}
\keyword{datasets}

