% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lightresponse}
\alias{lightresponse}
\title{Non-rectangular hyperbola}
\usage{
lightresponse(PAR, Amax, phi, theta, Rd, ...)
}
\arguments{
\item{PAR}{Photosynthetically active radiation (mu mol m-2 s-1)}

\item{Amax}{Maximum assimilation rate (the asymptote) (mu mol m-2 s-1)}

\item{phi}{Quantum yield (slope at PAR = 0) (mol mol-1)}

\item{theta}{Shape of light response curve (0 = rectangular hyperbola, 1 =
'Blackman' response).}

\item{Rd}{Dark respiration (**positive** value).}

\item{\dots}{Further arguments are ignored.}
}
\value{
Returns a dataframe with : \describe{ \item{A}{Net assimilation rate
(mu mol m-2 s-1)} }
}
\description{
A simple light response function that predicts leaf photosynthesis from
absorbed PAR.
}
\author{
Remko Duursma
}
\seealso{
\code{\link{setPhy}}
}
\keyword{misc}

