\name{makeStand}
\alias{makeStand}
\title{
Make a stand of virtual plants
}
\description{
empty
}
\usage{
makeStand(plants = list(), xyz = data.frame(x = 0, y = 0, z = 0), plotbox = NULL)
}

\arguments{
  \item{plants}{}
  \item{xyz}{}
  \item{plotbox}{}
}
\details{}

\value{}
\references{

}
\author{
Remko Duursma
}

%\examples{}
\keyword{misc}
