\name{plantexamples}
\alias{plantexamples}
\alias{toona}
\alias{sugarmaple}
\alias{pilularis}
\docType{data}
\title{
Three example 3D plants
}
\description{
Three virtual plant, in \code{YplantQMC} format. Objects of this kind would normally be produced
with \code{\link{constructplant}}.
}

\source{
Plants provided by Jeff Kelly and Kerrie Sendall.
}

\examples{

\dontrun{
# A sugar maple (Acer saccharum)
plot(sugarmaple)

# A blackbutt (Eucalyptus pilularis)
plot(pilularis)

# Australian red cedar (Toona australis)
plot(toona)
}

}
\keyword{datasets}
