% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrdata.R
\name{psrdata}
\alias{psrdata}
\title{Get PSR data.}
\usage{
psrdata(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link{YplantDay}}}

\item{\dots}{Further arguments ignored}
}
\description{
Extract Plant Summary Report from a YplantQMC simulation.

Simple wrapper: \code{x$psrdata} is the same as \code{psrdata(x).}
}
\author{
Remko Duursma
}
\keyword{misc}

