% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{turtle}
\alias{turtle}
\title{A turtle sky with 58 points}
\format{A data frame with 59 observations on the following 2 variables.
\describe{ \item{altitude}{a numeric vector}
\item{azimuth}{a numeric vector} }}
\description{
These are the angles used in \code{\link{STARbar}} when \code{integration =
"Turtlesky"}.
}
\keyword{datasets}

