% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{turtle482}
\alias{turtle482}
\title{A turtle sky with 482 points}
\format{A data frame with 482 observations on the following 2 variables.
\describe{
\item{altitude}{a numeric vector}
\item{azimuth}{a numeric vector} }}
\description{
A set of points that are uniformly distributed across the hemisphere, so that
each point represents (approx.) the same solid angle. Used in the diffuse
radiation calculations in \code{\link{YplantDay}}.
}
\keyword{datasets}

