\name{yplantaltaz}
\alias{yplantaltaz}
\docType{data}
\title{
Altitude and azimuth angles
}
\description{
The 160 angles used in the original Yplant (Pearcy & Yang 1996), in radians.
Used in \code{\link{STARbar}} when \code{integration = "Yplant"}.
}
\format{
  A data frame with 160 observations on the following 2 variables.
  \describe{
    \item{\code{azimuth}}{Azimuth angle (radians)}
    \item{\code{altitude}}{Altitude (radians)}
  }
}
\keyword{datasets}
