% Generated by roxygen2 (4.0.2): do not edit by hand
\name{installQuasiMC}
\alias{installQuasiMC}
\title{Download and install the QuasiMC executable}
\usage{
installQuasiMC(proxy = FALSE)
}
\arguments{
\item{proxy}{Use TRUE if connecting through a proxy. This invokes
\code{setInternet2} - only works on Windows.}
}
\description{
Downloads the Windows/Mac executable necessary for running
\code{\link{runYplant}} and \code{\link{YplantDay}}.

QuasiMC is developed by Cieslak et al. (2008), which is part of the Virtual
Laboratory plant modelling software created at the University of Calgary
(\url{http://www.algorithmicbotany.org}).

Both files are placed in the directory (for windows, \code{c:/QuasiMC}),
which is created if it does not exist already.

This function can also be used to update an existing installation.
}
\author{
QuasiMC by Mik Cieslak, R installation code by Remko Duursma.
}
\references{
Cieslak, M., C. Lemieux, J. Hanan and P. Prusinkiewicz. 2008.
Quasi-Monte Carlo simulation of the light environment of plants. Functional
Plant Biology. 35:837-849.

See also \url{http://www.remkoduursma/yplantqmc}
}
\keyword{misc}

