% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{plantexamples}
\alias{pilularis}
\alias{plantexamples}
\alias{sugarmaple}
\alias{toona}
\title{Three example 3D plants}
\source{
Plants provided by Jeff Kelly and Kerrie Sendall.
}
\description{
Three virtual plant, in \code{YplantQMC} format. Objects of this kind would
normally be produced with \code{\link{constructplant}}.
}
\examples{
\dontrun{
# A sugar maple (Acer saccharum)
plot(sugarmaple)

# A blackbutt (Eucalyptus pilularis)
plot(pilularis)

# Australian red cedar (Toona australis)
plot(toona)
}
}
\keyword{datasets}

