\name{psrdata}
\alias{psrdata}

\title{
Get PSR data.
}
\description{
Extract Plant Summary Report from a YplantQMC simulation.
}
\usage{
psrdata(x, ...)
}
\arguments{
  \item{x}{Object returned by \code{\link{YplantDay}}}
  \item{\dots}{Further arguments ignored}
}
\details{
Simple wrapper: \code{x$psrdata} is the same as \code{psrdata(x).}
}
% \value{
% %%  ~Describe the value returned
% %%  If it is a LIST, use
% %%  \item{comp1 }{Description of 'comp1'}
% %%  \item{comp2 }{Description of 'comp2'}
% %% ...
% % }
% \references{
% %% ~put references to the literature/web site here ~
% }
\author{
Remko Duursma
}
% \note{
% %%  ~~further notes~~
% }



\keyword{misc}

