% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yplantqmc-package.R
\docType{data}
\name{turtle244}
\alias{turtle244}
\title{A turtle sky with 244 points}
\format{A data frame with 244 observations on the following 2 variables.
\describe{ 
\item{altitude}{a numeric vector} 
\item{azimuth}{a numeric vector} }}
\description{
Not currently used.
}
\keyword{datasets}

