% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xmastime}
\alias{xmastime}
\title{Is it xmas yet?}
\usage{
xmastime(plant, palette = c("red", "blue2", "yellow", "darkorchid3", "gold1"),
  zadj = 1.5, nballsfrac = 0.2, ballradiusrel = 0.25)
}
\arguments{
\item{plant}{A 'plant3d' object (see \code{\link{constructplant}}).}

\item{palette}{Colors.}

\item{zadj}{Adjust relative height to leaf height.}

\item{nballsfrac}{Fraction of leaves with ornaments}

\item{ballradiusrel}{Relative radius of ornaments}
}
\description{
Prettify your plant.
}
\examples{
\dontrun{
plot(sugarmaple)
}
}
\author{
Santa
}
\keyword{misc}

