% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{yplantaltaz}
\alias{yplantaltaz}
\title{Altitude and azimuth angles}
\format{A data frame with 160 observations on the following 2 variables.
\describe{
\item{azimuth}{Azimuth angle (radians)}
\item{altitude}{Altitude (radians)} }}
\description{
The 160 angles used in the original Yplant (Pearcy & Yang 1996), in radians.
Used in \code{\link{STARbar}} when \code{integration = "Yplant"}.
}
\keyword{datasets}

